refactor: remove SqlCacheRepo, CacheRepo, models-store helpers
Every former reader has been migrated to a typed repo (modelsCache,
codexPkcePending, searchConfig, image_cache via ImageCacheStore) or to
per-provider state_json. The legacy `cache: CacheRepo` field on Repo and
ProviderRepo is gone, the SqlCacheRepo/MemoryCacheRepo impls are
deleted, and models-store.ts is renamed to models-fetch.ts with the
cache helpers stripped (keeping only the fetch helper + error types).
tsc confirms no readers remain.

Test-only consequences:
- The legacy gateway models-store_test.ts is deleted: every helper it
  exercised (inProcessMemo / readModelsStore / writeModelsStore /
  invalidateModelsStore / clearModelsStore) is gone, and the SWR cache
  is covered by models-cache_test.ts.
- The four "Custom provider serves from L2 / re-fetches / reuses stale
  on 429 / throws beyond hard window" tests in custom/provider_test.ts
  asserted in-process memo + repo cache behaviour around
  createCustomProvider().getProvidedModels(); that layer no longer
  caches at all (the SWR cache lives in fetchUpstreamModelsCached), so
  the tests were exercising dead code paths and have been removed.
- isProviderModelsHttpStatus had no production caller; the helper and
  its provider-custom test asserts are dropped.
- test-utils' memoryCacheRepo factory is removed; provider-codex /
  provider-copilot test stubs no longer install a cache field.

Drive-by lint fixes flagged in earlier task reviews:
- Trailing comma on the codex_pkce_pending INSERT in SqlRepo.
- Closing-paren placement in createCodexProvider getProvidedModels.
