+++ packages/gateway/src/data-plane/embeddings/serve_test.ts
@@ -1,7 +1,6 @@
 import { test } from 'vitest';
 
 import { buildCustomUpstreamRecord, copilotModels, flushAsyncWork, requestApp, setupAppTest } from '../../test-helpers.ts';
-import { clearModelsStore } from '@floway-dev/provider';
 import { clearCopilotTokenCache } from '@floway-dev/provider-copilot';
 import { jsonResponse, withMockedFetch, assertEquals, assertExists } from '@floway-dev/test-utils';
 
@@ -203,7 +202,6 @@ test('/v1/embeddings records request and upstream performance', async () => {
 test('/v1/embeddings routes to custom upstream when model is only declared there', async () => {
   const { apiKey, repo } = await setupAppTest();
   await repo.upstreams.deleteAll();
-  clearModelsStore();
   await clearCopilotTokenCache();
 
   await repo.upstreams.save(buildCustomUpstreamRecord({
@@ -276,7 +274,6 @@ test('/v1/embeddings routes to custom upstream when model is only declared there
 test('/v1/embeddings rejects model on custom upstream without /embeddings capability', async () => {
   const { apiKey, repo } = await setupAppTest();
   await repo.upstreams.deleteAll();
-  clearModelsStore();
   await clearCopilotTokenCache();
 
   await repo.upstreams.save(buildCustomUpstreamRecord({
@@ -330,7 +327,6 @@ test('/v1/embeddings rejects model on custom upstream without /embeddings capabi
 test('/v1/embeddings preserves custom upstream /models HTTP errors', async () => {
   const { apiKey, repo } = await setupAppTest();
   await repo.upstreams.deleteAll();
-  clearModelsStore();
   await clearCopilotTokenCache();
 
   await repo.upstreams.save(buildCustomUpstreamRecord({
@@ -381,7 +377,6 @@ test('/v1/embeddings preserves custom upstream /models HTTP errors', async () =>
 
 test('/v1/embeddings preserves model-load errors hidden by another provider', async () => {
   const { apiKey, repo } = await setupAppTest();
-  clearModelsStore();
   await clearCopilotTokenCache();
 
   await repo.upstreams.save(buildCustomUpstreamRecord({
