+++ packages/gateway/src/data-plane/images/serve_test.ts
@@ -1,7 +1,6 @@
 import { test } from 'vitest';
 
 import { buildCustomUpstreamRecord, copilotModels, flushAsyncWork, requestApp, setupAppTest } from '../../test-helpers.ts';
-import { clearModelsStore } from '@floway-dev/provider';
 import { clearCopilotTokenCache } from '@floway-dev/provider-copilot';
 import { jsonResponse, withMockedFetch, assertEquals, assertExists } from '@floway-dev/test-utils';
 
@@ -76,7 +75,6 @@ test('/v1/images/edits rejects multipart body without model field with 400', asy
 test('/v1/images/generations rejects model on custom upstream without /images/generations capability', async () => {
   const { apiKey, repo } = await setupAppTest();
   await repo.upstreams.deleteAll();
-  clearModelsStore();
   await clearCopilotTokenCache();
 
   // Chat-only custom upstream. Its /models response advertises gpt-4o
@@ -118,7 +116,6 @@ test('/v1/images/generations forwards a JSON request through a custom upstream a
   const { apiKey, repo } = await setupAppTest();
   // setupAppTest seeds only a Copilot upstream by default; register a custom
   // upstream whose /models response declares the requested image model.
-  clearModelsStore();
   await clearCopilotTokenCache();
   await repo.upstreams.save(buildCustomUpstreamRecord({
     id: 'up_images',
@@ -174,7 +171,6 @@ test('/v1/images/edits forwards a multipart request through an Azure model and r
   // setupAppTest seeds a Copilot upstream by default; we also register an
   // Azure upstream that exposes gpt-image-2 via /v1/images/edits.
   const { apiKey, repo } = await setupAppTest();
-  clearModelsStore();
   await clearCopilotTokenCache();
   await repo.upstreams.save({
     id: 'az-image',
