+++ packages/gateway/src/data-plane/models/gemini_test.ts
@@ -1,7 +1,6 @@
 import { test } from 'vitest';
 
 import { buildCustomUpstreamRecord, copilotModels, requestApp, setupAppTest } from '../../test-helpers.ts';
-import { clearModelsStore } from '@floway-dev/provider';
 import { clearCopilotTokenCache } from '@floway-dev/provider-copilot';
 import { jsonResponse, withMockedFetch, assertEquals } from '@floway-dev/test-utils';
 
@@ -135,7 +134,6 @@ test('/v1beta/models/:modelId returns one Gemini model or Google RPC 404', async
 test('/v1beta/models includes custom upstream LLM models', async () => {
   const { apiKey, repo } = await setupAppTest();
   await repo.upstreams.deleteAll();
-  clearModelsStore();
   await clearCopilotTokenCache();
 
   await repo.upstreams.save(buildCustomUpstreamRecord({
@@ -187,7 +185,6 @@ test('/v1beta/models includes custom upstream LLM models', async () => {
 test('/v1beta/models excludes custom upstream embedding-only models', async () => {
   const { apiKey, repo } = await setupAppTest();
   await repo.upstreams.deleteAll();
-  clearModelsStore();
   await clearCopilotTokenCache();
 
   await repo.upstreams.save(buildCustomUpstreamRecord({
@@ -229,7 +226,6 @@ test('/v1beta/models excludes custom upstream embedding-only models', async () =
 test('/v1beta/models hides upstream identity when a provider returns an invalid model list', async () => {
   const { apiKey, repo } = await setupAppTest();
   await repo.upstreams.deleteAll();
-  clearModelsStore();
   await clearCopilotTokenCache();
 
   await repo.upstreams.save(buildCustomUpstreamRecord({
@@ -273,7 +269,6 @@ test('/v1beta/models hides upstream identity when a provider returns an invalid 
 test('/v1beta/models hides upstream HTTP error bodies', async () => {
   const { apiKey, repo } = await setupAppTest();
   await repo.upstreams.deleteAll();
-  clearModelsStore();
   await clearCopilotTokenCache();
 
   await repo.upstreams.save(buildCustomUpstreamRecord({
@@ -320,7 +315,6 @@ test('/v1beta/models hides upstream HTTP error bodies', async () => {
 test('/v1beta/models hides thrown upstream request errors', async () => {
   const { apiKey, repo } = await setupAppTest();
   await repo.upstreams.deleteAll();
-  clearModelsStore();
   await clearCopilotTokenCache();
 
   await repo.upstreams.save(buildCustomUpstreamRecord({
@@ -364,7 +358,6 @@ test('/v1beta/models hides thrown upstream request errors', async () => {
 test('/v1beta/models hides malformed upstream response bodies', async () => {
   const { apiKey, repo } = await setupAppTest();
   await repo.upstreams.deleteAll();
-  clearModelsStore();
   await clearCopilotTokenCache();
 
   await repo.upstreams.save(buildCustomUpstreamRecord({
