+++ packages/gateway/src/data-plane/models/serve_test.ts
@@ -1,7 +1,6 @@
 import { test } from 'vitest';
 
 import { buildCopilotUpstreamRecord, buildCustomUpstreamRecord, copilotModels, requestApp, setupAppTest } from '../../test-helpers.ts';
-import { clearModelsStore } from '@floway-dev/provider';
 import { clearCopilotTokenCache } from '@floway-dev/provider-copilot';
 import { jsonResponse, withMockedFetch, assertEquals } from '@floway-dev/test-utils';
 
@@ -276,7 +275,6 @@ test('/models returns the same superset payload as /v1/models', async () => {
 test('/v1/models hides upstream identity when a provider returns an invalid model list', async () => {
   const { repo, apiKey } = await setupAppTest();
   await repo.upstreams.deleteAll();
-  clearModelsStore();
   await clearCopilotTokenCache();
   await repo.upstreams.save(buildCustomUpstreamRecord({
     id: 'up_secret_provider',
@@ -312,7 +310,6 @@ test('/v1/models hides upstream identity when a provider returns an invalid mode
 test('public model list endpoints hide upstream HTTP error bodies and headers', async () => {
   const { repo, apiKey } = await setupAppTest();
   await repo.upstreams.deleteAll();
-  clearModelsStore();
   await clearCopilotTokenCache();
   await repo.upstreams.save(buildCustomUpstreamRecord({
     id: 'up_http_secret_provider',
@@ -360,7 +357,6 @@ test('public model list endpoints hide upstream HTTP error bodies and headers', 
 test('public model list endpoints hide thrown upstream request errors', async () => {
   const { repo, apiKey } = await setupAppTest();
   await repo.upstreams.deleteAll();
-  clearModelsStore();
   await clearCopilotTokenCache();
   await repo.upstreams.save(buildCustomUpstreamRecord({
     id: 'up_throw_secret_provider',
@@ -401,7 +397,6 @@ test('public model list endpoints hide thrown upstream request errors', async ()
 test('public model list endpoints hide malformed upstream response bodies', async () => {
   const { repo, apiKey } = await setupAppTest();
   await repo.upstreams.deleteAll();
-  clearModelsStore();
   await clearCopilotTokenCache();
   await repo.upstreams.save(buildCustomUpstreamRecord({
     id: 'up_malformed_secret_provider',
@@ -445,7 +440,6 @@ test('public model list endpoints hide malformed upstream response bodies', asyn
 test('/v1/models surfaces the actionable "no upstream configured" hint when no provider is configured', async () => {
   const { repo, apiKey } = await setupAppTest();
   await repo.upstreams.deleteAll();
-  clearModelsStore();
   await clearCopilotTokenCache();
 
   const response = await requestApp('/v1/models', {
