+++ packages/provider/src/repo.ts
@@ -1,16 +1,5 @@
 import type { UpstreamRecord } from './model.ts';
 
-// Subset of the gateway-internal Repo that provider-layer code actually reads
-// from. packages/gateway wires its concrete repo accessor at boot via
-// `initProviderRepo` so provider-package helpers never reach back into the
-// gateway.
-export interface CacheRepo {
-  get(key: string): Promise<string | null>;
-  set(key: string, value: string, ttlMs?: number): Promise<void>;
-  delete(key: string): Promise<void>;
-  deletePrefix(prefix: string): Promise<void>;
-}
-
 // Slim upstream-state surface for providers that own autonomous runtime state
 // (e.g. Codex's rotated tokens). Structurally compatible with the full
 // UpstreamRepo in packages/gateway, so the wiring stays a single accessor.
@@ -20,7 +9,6 @@ export interface UpstreamsRepoSlim {
 }
 
 export interface ProviderRepo {
-  cache: CacheRepo;
   upstreams: UpstreamsRepoSlim;
 }
