+++ packages/test-utils/src/index.ts
@@ -9,4 +9,4 @@ export {
   assertThrows,
 } from './assert.ts';
 export { jsonResponse, sseResponse, withMockedFetch } from './mock-fetch.ts';
-export { memoryCacheRepo, noopUpstreamCallOptions, stubProvider, stubProviderCandidate, stubUpstreamModel, testTelemetryModelIdentity } from './stubs.ts';
+export { noopUpstreamCallOptions, stubProvider, stubProviderCandidate, stubUpstreamModel, testTelemetryModelIdentity } from './stubs.ts';
